# [opt](arm)Remove negative optimizations of SSE2NEON on memcmp for ARM